Switch the Gym automated logger setup to configure the root logger rather than just the 'gym' logger
Python doesn't make it easy for libraries to take responsibility for
logging configuration (which we do to make simple usage much easier),
and as we see more Gym plugins, we want their loggers to have an
appropriate log level too. So we may as well configure the root logger
level.