Unbreak random_agent
The semantics of `env.reset()` have changed. It's no longer possible
to reset an environment until it's done.

Notably, this commit changes the behavior of random_agent. It used
to have a limit on number of actions per episode. If we wanted
that now, we'd have to close the environment and recreate it
on each episode, which may be slow.