save/load Atari environment state
save (clone) and load (restore) the Atari emulator state for analysis,
planning, or other experiments that require returning to identical
states of the environment.

note: `clone_full_state()` and `restore_full_state()` are necessary to
save and load the ROM and system state (including pseudorandomness) for
fully identical restoration of the environment.