Add gym-jiminy in the third-party environments doc
I have developed a new open source simulator for poly-articulated system called Jiminy. It relies on the powerful rigid body dynamics framework Pinocchio for the physics evaluation, and meshcat for real-time visualization.
I think the project is now mature enough for everyone to use it, despite being partially undocumented. The library has been designed for fast prototyping of control algorithms and machine learning. It is easy to use and several examples of classical control theory and machine learning are available (cartpole, double and simple pendulums).