Set default image color to white
In gym.envs.classic_control.rendering, Geom has a color attribute,
which is initialized as black. This is a good default for most Geom
subclasses, such as lines and circles. However, in the case of Image
instances, the result is a completely black image - in the case of an
image without transparency, the result is a black square. This can be
easily fixed by setting the default color for Image instances as white,
which causes images to be displayed normally.