Populate TimeLimit.truncated on every step
Writing samplers which use the `info` return from `Env.step()` is much easier if the keys in `info` are stable.

This PR populates `info['TimeLimit.truncated']` on every step of training, not just on truncated steps. This allows samplers to test for timeouts without having to test for key existence first.